move providers to root
