+++ app/app/(root)/layout.tsx
@@ -1,20 +1,20 @@
-import { NavBar } from '@/components/nav-bar'
-import '@/styles/globals.css'
-import type { Metadata } from 'next'
-import { Inter } from 'next/font/google'
-import { Providers } from '../providers'
+import { NavBar } from "@/components/nav-bar";
+import "@/styles/globals.css";
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import { Providers } from "../../providers";
 
-const inter = Inter({ subsets: ['latin'] })
+const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: 'Portal',
-  description: 'Portal',
-}
+  title: "Portal",
+  description: "Portal",
+};
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
     <html lang="en">
@@ -27,5 +27,5 @@ export default function RootLayout({
         </Providers>
       </body>
     </html>
-  )
+  );
 }
